backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
lxml==5.3.0
trimesh==4.4.9
numpy==1.26.4
shapely>=2.0.0
opencv-python-headless==4.8.1.78
Pillow>=9.0.0
scikit-image>=0.19.0
